Move some code out of `tools/building.py`. NFC
Some of these function are only called in single file so
there is no need for them to be in this shared location.
