+++ test/test_other.py
@@ -58,6 +58,24 @@ emsymbolizer = shared.bat_suffix(path_from_root('emsymbolizer'))
 wasm_opt = Path(building.get_binaryen_bin(), 'wasm-opt')
 
 
+def is_bitcode(filename):
+  try:
+    # look for magic signature
+    b = open(filename, 'rb').read(4)
+    if b[:2] == b'BC':
+      return True
+    # on macOS, there is a 20-byte prefix which starts with little endian
+    # encoding of 0x0B17C0DE
+    elif b == b'\xDE\xC0\x17\x0B':
+      b = bytearray(open(filename, 'rb').read(22))
+      return b[20:] == b'BC'
+  except IndexError:
+    # not enough characters in the input
+    # note that logging will be done on the caller function
+    pass
+  return False
+
+
 def uses_canonical_tmp(func):
   """Decorator that signals the use of the canonical temp by a test method.
 
@@ -5048,7 +5066,7 @@ EM_ASM({ _middle() });
     self.run_process([EMCC, '-flto', '-c', test_file('hello_world.c')])
     self.assertExists('hello_world.o')
     self.run_process([EMCC, '-flto', '-r', 'hello_world.o', '-o', 'hello_world2.o'])
-    building.is_bitcode('hello_world.o')
+    is_bitcode('hello_world.o')
     building.is_wasm('hello_world2.o')
 
   @parameterized({
@@ -5588,7 +5606,7 @@ int main() {
 
   def test_emit_llvm(self):
     self.run_process([EMCC, test_file('hello_world.c'), '-c', '-emit-llvm'])
-    self.assertTrue(building.is_bitcode('hello_world.bc'))
+    self.assertTrue(is_bitcode('hello_world.bc'))
 
   def test_compile_ll_file(self):
     self.run_process([EMCC, test_file('hello_world.c'), '-S', '-emit-llvm'])
@@ -8775,12 +8793,12 @@ int main() {
       print('wasm in object')
       self.run_process([EMXX, src] + args + ['-c', '-o', 'hello_obj.o'])
       self.assertTrue(building.is_wasm('hello_obj.o'))
-      self.assertFalse(building.is_bitcode('hello_obj.o'))
+      self.assertFalse(is_bitcode('hello_obj.o'))
 
       print('bitcode in object')
       self.run_process([EMXX, src] + args + ['-c', '-o', 'hello_bitcode.o', '-flto'])
       self.assertFalse(building.is_wasm('hello_bitcode.o'))
-      self.assertTrue(building.is_bitcode('hello_bitcode.o'))
+      self.assertTrue(is_bitcode('hello_bitcode.o'))
 
       print('use bitcode object (LTO)')
       self.run_process([EMXX, 'hello_bitcode.o'] + args + ['-flto'])
@@ -8813,7 +8831,7 @@ int main() {
       (['-sWASM_OBJECT_FILES'], False),
     ]:
       self.run_process([EMCC, test_file('hello_world.c')] + flags + ['-c', '-o', 'a.o'])
-      seen_bitcode = building.is_bitcode('a.o')
+      seen_bitcode = is_bitcode('a.o')
       self.assertEqual(expect_bitcode, seen_bitcode, 'must emit LTO-capable bitcode when flags indicate so (%s)' % str(flags))
 
   # We have LTO tests covered in 'wasmltoN' targets in test_core.py, but they
@@ -10181,17 +10199,17 @@ test_module().then((test_module_instance) => {
 
     with open(fname, 'wb') as f:
       f.write(b'foo')
-    self.assertFalse(building.is_bitcode(fname))
+    self.assertFalse(is_bitcode(fname))
 
     with open(fname, 'wb') as f:
       f.write(b'\xDE\xC0\x17\x0B')
       f.write(16 * b'\x00')
       f.write(b'BC')
-    self.assertTrue(building.is_bitcode(fname))
+    self.assertTrue(is_bitcode(fname))
 
     with open(fname, 'wb') as f:
       f.write(b'BC')
-    self.assertTrue(building.is_bitcode(fname))
+    self.assertTrue(is_bitcode(fname))
 
   def test_is_ar(self):
     fname = 'tmp.a'
@@ -11987,7 +12005,7 @@ Aborted(`Module.arguments` has been replaced by `arguments_` (the initial value 
     # Verify that bitcode files are accepted as input
     create_file('main.c', 'void foo(); int main() { return 0; }')
     self.run_process([EMCC, '-emit-llvm', '-c', '-o', 'main.bc', 'main.c'])
-    self.assertTrue(building.is_bitcode('main.bc'))
+    self.assertTrue(is_bitcode('main.bc'))
     self.run_process([EMCC, '-c', '-o', 'main.o', 'main.bc'])
     self.assertTrue(building.is_wasm('main.o'))
