+++ tools/link.py
@@ -1978,6 +1978,20 @@ def phase_embind_aot(wasm_target, js_syms):
   write_file(final_js, src)
 
 
+# for Popen, we cannot have doublequotes, so provide functionality to
+# remove them when needed.
+def remove_quotes(arg):
+  if isinstance(arg, list):
+    return [remove_quotes(a) for a in arg]
+
+  if arg.startswith('"') and arg.endswith('"'):
+    return arg[1:-1].replace('\\"', '"')
+  elif arg.startswith("'") and arg.endswith("'"):
+    return arg[1:-1].replace("\\'", "'")
+  else:
+    return arg
+
+
 @ToolchainProfiler.profile_block('source transforms')
 def phase_source_transforms(options):
   # Apply a source code transformation, if requested
@@ -1986,7 +2000,7 @@ def phase_source_transforms(options):
   final_js += '.tr.js'
   posix = not shared.WINDOWS
   logger.debug('applying transform: %s', options.js_transform)
-  shared.check_call(building.remove_quotes(shlex.split(options.js_transform, posix=posix) + [os.path.abspath(final_js)]))
+  shared.check_call(remove_quotes(shlex.split(options.js_transform, posix=posix) + [os.path.abspath(final_js)]))
   save_intermediate('transformed')
 
 
@@ -2637,6 +2651,85 @@ def find_library(lib, lib_dirs):
   return None
 
 
+def map_to_js_libs(library_name):
+  """Given the name of a special Emscripten-implemented system library, returns an
+  pair containing
+  1. Array of absolute paths to JS library files, inside emscripten/src/ that corresponds to the
+     library name. `None` means there is no mapping and the library will be processed by the linker
+     as a require for normal native library.
+  2. Optional name of a corresponding native library to link in.
+  """
+  # Some native libraries are implemented in Emscripten as system side JS libraries
+  library_map = {
+    'embind': ['embind/embind.js', 'embind/emval.js'],
+    'EGL': ['library_egl.js'],
+    'GL': ['library_webgl.js', 'library_html5_webgl.js'],
+    'webgl.js': ['library_webgl.js', 'library_html5_webgl.js'],
+    'GLESv2': ['library_webgl.js'],
+    # N.b. there is no GLESv3 to link to (note [f] in https://www.khronos.org/registry/implementers_guide.html)
+    'GLEW': ['library_glew.js'],
+    'glfw': ['library_glfw.js'],
+    'glfw3': ['library_glfw.js'],
+    'GLU': [],
+    'glut': ['library_glut.js'],
+    'openal': ['library_openal.js'],
+    'X11': ['library_xlib.js'],
+    'SDL': ['library_sdl.js'],
+    'uuid': ['library_uuid.js'],
+    'fetch': ['library_fetch.js'],
+    'websocket': ['library_websocket.js'],
+    # These 4 libraries are separate under glibc but are all rolled into
+    # libc with musl.  For compatibility with glibc we just ignore them
+    # completely.
+    'dl': [],
+    'm': [],
+    'rt': [],
+    'pthread': [],
+    # This is the name of GNU's C++ standard library. We ignore it here
+    # for compatibility with GNU toolchains.
+    'stdc++': [],
+  }
+  settings_map = {
+    'glfw': {'USE_GLFW': 2},
+    'glfw3': {'USE_GLFW': 3},
+    'SDL': {'USE_SDL': 1},
+  }
+
+  if library_name in settings_map:
+    for key, value in settings_map[library_name].items():
+      default_setting(key, value)
+
+  if library_name in library_map:
+    libs = library_map[library_name]
+    logger.debug('Mapping library `%s` to JS libraries: %s' % (library_name, libs))
+    return libs
+
+  if library_name.endswith('.js') and os.path.isfile(utils.path_from_root('src', f'library_{library_name}')):
+    return [f'library_{library_name}']
+
+  return None
+
+
+# Map a linker flag to a settings. This lets a user write -lSDL2 and it will
+# have the same effect as -sUSE_SDL=2.
+def map_and_apply_to_settings(library_name):
+  # most libraries just work, because the -l name matches the name of the
+  # library we build. however, if a library has variations, which cause us to
+  # build multiple versions with multiple names, then we need this mechanism.
+  library_map = {
+    # SDL2_mixer's built library name contains the specific codecs built in.
+    'SDL2_mixer': [('USE_SDL_MIXER', 2)],
+  }
+
+  if library_name in library_map:
+    for key, value in library_map[library_name]:
+      logger.debug('Mapping library `%s` to settings changes: %s = %s' % (library_name, key, value))
+      setattr(settings, key, value)
+    return True
+
+  return False
+
+
 def process_libraries(state, linker_inputs):
   new_flags = []
   libraries = []
@@ -2652,7 +2745,7 @@ def process_libraries(state, linker_inputs):
 
     logger.debug('looking for library "%s"', lib)
 
-    js_libs = building.map_to_js_libs(lib)
+    js_libs = map_to_js_libs(lib)
     if js_libs is not None:
       libraries += [(i, js_lib) for js_lib in js_libs]
 
@@ -2667,7 +2760,7 @@ def process_libraries(state, linker_inputs):
     if js_libs is not None:
       continue
 
-    if building.map_and_apply_to_settings(lib):
+    if map_and_apply_to_settings(lib):
       continue
 
     path = None
